Refactor: List 와 ListItem 분리 후 이동
