+++ src/components/common/List.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import ListItem from './ListItem';
+
+const List = ({ data }) => {
+  return (
+    <section>
+      {data.map((review) => (
+        <ListItem review={review} />
+      ))}
+    </section>
+  );
+};
+
+export default List;
